split up app construction using targets
this allows a fork, where yarn or bundler packages can get updated
without one forcing the other into a rebuild.  it _also_ allows
installation of one or the other without requiring build dependencies to
go into the final image ... but I'm not sure if any of the current
depenency packages are actually optional, in the current case.
Something to poke around with.